requirements.txt
langchain
llama-index
crewai
chromadb
    
openai
transformers
sentence-transformers

fastapi
uvicorn
pydantic

sqlalchemy
psycopg2-binary
pyodbc
mysql-connector-python
snowflake-connector-python

tqdm
python-dotenv
loguru

pytest